requirements.txt
torch==2.1.0
transformers==4.42.3
accelerate==0.31.0
datasets==2.20.0
evaluate==0.4.2
pyyaml==6.0.1
scikit-learn==1.5.1
matplotlib==3.9.0
openai==1.35.9
seaborn==0.13.2
beautifulsoup4==4.12.3
lxml==5.2.2
anthropic==0.30.1
google-generativeai==0.7.2
tiktoken==0.7.0